requirements.txt
torch
torchvision
segmentation-models-pytorch==0.3.3
torchmetrics
albumentations
opencv-python
matplotlib
numpy
pillow
rembg
streamlit
